Eye in hand creation
The concept here is that the camera will be moving together with the robotic eye. As the robotic eye can only work in 2 Dimensions at the moment, So we have to provide states for the robot to understand where it is right now and the possible actions that it could perform.